Fix a rare bug that prevents user registration when $.mailcheck() was activated
This one would take a while to explain; Basically, $.mailcheck() will
prevent the first submit event in order to give the user a chance to
correct a wrongly-inputted email address, and will let the event fire
only the second time around. When combined with $.showLoading() however
the form would become inaccessible after the first submit event (blocked
by mailcheck) and the user will have no means of finishing the submit
process.

Every form that combined $.mailcheck() and $.showLoading() was affected.