Add cqWidgetFormNullInput for truly hidden form fields
The null widget is used when we need to have a form field that's availalbe
to the form, but that will never be outputted in a template (even as a
hidden field). For example, a hidden IP field used for spam validation.

The Bootstrap form formatter has also been modified to work properly with
the new widget.