Do not do DI for sfUser and ComposePrivateMessageForm
We were not addding the user to the Abridged PM form, so the check for
number of PMs set was never executed when messaging users from their pages.
Instead of correcting every instance of ComposeAbridgedPMForm, I simply
add the sf_user from the context to the form object in its consturctor.